% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_pca.R
\name{biplot_factoextra}
\alias{biplot_factoextra}
\title{Biplot with factorextra}
\usage{
biplot_factoextra(
  prcomp_element,
  groups = "none",
  ellipse = FALSE,
  loadings = "none",
  out_path = "none"
)
}
\arguments{
\item{prcomp_element}{object of class prcomp. The base::prcomp function performs a principal component analysis.
This is the object it returns.}

\item{groups}{vector. Groups to color by. Default = "none"}

\item{ellipse}{bool. Draw confidence ellipse arount the clusters of groups. Default = FALSE.}

\item{loadings}{string / vector of strings. Draw loadings arrows and labels. c("arrow", "text"), "text", "none" (default).}

\item{out_path}{string. Path to save plot to png.
If out_path is empty, the plot is printed to the device.}
}
\description{
`biplot_factoextra` prints a biplot from a prcomp element.
}
\details{
This function prints a biplot from a prcomp element.
The points can be colored by group. Loadings are optionally displayed.
}
\examples{
{pca_iris <- prcomp(dplyr::select_if(iris, is.numeric))
iris_groups <- iris$Species
biplot_factoextra(pca_iris)
biplot_factoextra(pca_iris,
                  groups = iris_groups,
                  ellipse = TRUE,
                  loadings = c("arrow", "text"))}
}
