% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_pca.R
\name{biplot_ggplot2}
\alias{biplot_ggplot2}
\title{Biplot with ggplot2}
\usage{
biplot_ggplot2(
  input_df,
  xdim = 1,
  ydim = 2,
  groups = "none",
  scale = FALSE,
  loadings = FALSE,
  ellipse = FALSE,
  title = "PCA - Biplot",
  out_path = "none"
)
}
\arguments{
\item{input_df}{a data frame with at least one factor column.}

\item{xdim}{integer. Default: 1}

\item{ydim}{integer. Default: 2}

\item{groups}{vector. Groups to color by. Default = "none"}

\item{scale}{logical. Should data be scaled to unit variance for PCA? Default = FALSE}

\item{loadings}{logical. Draw loadings arrows and labels. Default = FALSE}

\item{ellipse}{logical. Draw confidence ellipse arount the clusters of groups. Default = FALSE.}

\item{title}{string. Main title of the plot. Default: "PCA - Biplot"}

\item{out_path}{string. Path to save plot to png. If out_path is empty, the plot is printed to the device.}
}
\description{
`biplot_ggplot2` prints a biplot from a prcomp element.
}
\details{
This function prints a biplot from a prcomp element. The points can be colored by group. Loadings are optionally displayed.
}
\examples{
\dontrun{
biplot_ggplot2(iris, groups = "Species")
biplot_ggplot2(iris, groups = "Species", xdim = 1, ydim = 3  )
biplot_ggplot2(iris, "Species", ellipse = TRUE, loadings = TRUE)}
}
