% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_visualization.R
\name{boxplot_by_factor}
\alias{boxplot_by_factor}
\title{Print boxplot by factor}
\usage{
boxplot_by_factor(input_df, factor, out_path = "none")
}
\arguments{
\item{input_df}{a data frame with at least one factor column}

\item{factor}{a string with the column name to group by}

\item{out_path}{optional string.}
}
\description{
`boxplot_by_factor` generates plot with one boxplot per group for each variable
}
\details{
A plot with one boxplot per group is generated for all variables of a data frame.
}
\examples{
boxplot_by_factor(iris, "Species")
\dontrun{
dir.create(paste(getwd(), "/examples", sep = ""), showWarnings = FALSE)
dir <- paste(getwd(), "/examples/iris", sep = "")
boxplot_by_factor(iris, "Species", dir)
}

}
