% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_dataTransformaions.R
\name{calc_by_replicate}
\alias{calc_by_replicate}
\title{Aggregate function results by factor}
\usage{
calc_by_replicate(input_df, factor, funct, na_action = na.omit)
}
\arguments{
\item{input_df}{a data frame with at least one factor column}

\item{factor}{a string with the column name to group by}

\item{funct}{a generic R function, that takes only one argument (e.g. mean(), summary(), etc.)}

\item{na_action}{function. Indcates what should happen if there are NA values in the data. c(NULL (default), na.omit).}
}
\description{
`calc_by_replicate` takes a data frame calculates the results of a fuction grouped by a factor
}
\details{
Take a data frame and calculate the results of a fuction grouped by a given factor.
}
