% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_visualization.R
\name{calculate_ratio_matrix}
\alias{calculate_ratio_matrix}
\title{Calculate ratio matrix}
\usage{
calculate_ratio_matrix(input_vector, names_vector = 1:length(input_vector))
}
\arguments{
\item{input_vector}{numeric vector.}

\item{names_vector}{vector. Length = length(input_vector). Row and column names for the returnded matrix. Default are numbers 1:length(input_vector).}
}
\description{
`calculate_ratio_matrix` takes a vector and calculates the ratio between all elements of a vector.
}
\examples{
myvector <- c(1712.9583, 1446.4583, 1968.4167, 2124.1250, 2315.7083, 1135.1667, 2227.5000)
mynames <- letters[1:length(myvector)]
calculate_ratio_matrix(myvector, mynames)
}
