% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_visualization.R
\name{correlation_plot}
\alias{correlation_plot}
\title{Correlation plots}
\usage{
correlation_plot(input_df, method = "spearman", out_path = "none")
}
\arguments{
\item{input_df}{data frame.}

\item{method}{string. Method for calculating the correlation. Options: "pearson", "kendall", "spearman" (default).}

\item{out_path}{optional string. Path to save correlation plot to png. If out_path is empty the correlation plot is printed to the device.}
}
\description{
`correlation_plot` calculates correlations of variables and displays them in a dotplot
}
\details{
A grid of dotplots displaying the correlations between the variables of a data frame is generated. Use with less than 10 variables ideally.
}
\examples{
correlation_plot(iris)
\dontrun{
dir.create(paste(getwd(), "/examples", sep = ""), showWarnings = FALSE)
dir <- paste(getwd(), "/examples/iris", sep = "")
correlation_plot(iris, out_path = dir)
}
}
