% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_read_TF.R
\name{flip_df}
\alias{flip_df}
\title{Flip data frame}
\usage{
flip_df(df)
}
\arguments{
\item{df}{data frame. Contains the columns c(Compound, Type, Filename, Status, Group, Area).}
}
\description{
`flip_df` takes a data frame with the columns c(Compound, Type, Filename, Status, Group, Area) and transposes it, so that each area value is assigned to one compound (columnwise) and one SID (row-wise). Additionally the group variable is saved to each sample.  Suitable for preprocessed csvs generated by Thermo Fisher Tracefinder.
}
\examples{
flip_df(tfData)
}
