% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_dataTransformaions.R
\name{generate_categorical_table}
\alias{generate_categorical_table}
\title{Generate table of categorical variable}
\usage{
generate_categorical_table(categorical_col)
}
\arguments{
\item{categorical_col}{column of data frame containing categorical values (i.e. strings). Format = df$colname}
}
\description{
`generate_categorical_table` takes a categorical column of a data frame and generates a table where the categorical strings are assigned to a numeric code. This way the information of additional catecorical variabes is not lost when performing an aggregate function.
}
