% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_clustering.R
\name{hclust_dendrogram}
\alias{hclust_dendrogram}
\title{Hierarchical clustering dendrogram}
\usage{
hclust_dendrogram(
  hclust_element,
  labs = hclust_element$labels,
  title = "Hierarchical clustering",
  out_path = "none"
)
}
\arguments{
\item{hclust_element}{object of class hclust.}

\item{labs}{vector. Labels of dendrogram. Default: hclust_element$labels}

\item{title}{string. Plot title. Default = "Hierarchical clustering".}

\item{out_path}{string. Path to save plot to png. Default = to device.}
}
\description{
`hclust_dendrogram` takes a data frame and prints a dendrogram.
}
\details{
This function takes a data frame and transforms it into a matrix. The given matrix is then used to perform hiererchical clustering with the average, single or complete linkage. The clustering is plotted in a dendrogram.
}
\examples{
US_clust <- stats::hclust(stats::dist(USArrests, method = "euclidean"), method = "average")
hclust_dendrogram(US_clust)
hclust_dendrogram(US_clust, labs = 1:nrow(USArrests))
\dontrun{
dir.create(paste(getwd(), "/examples", sep = ""), showWarnings = FALSE)
dir <- paste(getwd(), "/examples/USArrests", sep = "")
hclust_dendrogram(US_clust, out_path = dir)}
}
