% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_clustering.R
\name{hclust_heatmap}
\alias{hclust_heatmap}
\title{Hierarchical clustering heatmap}
\usage{
hclust_heatmap(
  input_df,
  row_names = rownames(input_df),
  col_names = colnames(dplyr::select_if(input_df, is.numeric)),
  dist_method = "euclidean",
  hclust_method = "complete",
  title = "Hierarchical clustering",
  out_path = "none"
)
}
\arguments{
\item{input_df}{data frame.}

\item{row_names}{vector of length nrow(input_df). Default = rownames(input_df).}

\item{col_names}{vector of length ncol(select_if(input_df, is.numeric)). Default = colnames(select_if(input_df, is.numeric))}

\item{dist_method}{string. Method for the dist() function. dist_method = c("euclidean", "manhattan"). Default = "euclidean".}

\item{hclust_method}{string. Method for the hclust() function. hclust_method = c("complete", "single", "average"). Default = "complete".}

\item{title}{string. Plot title. Default = "Hierarchical clustering".}

\item{out_path}{string. Path to save plot to png.}
}
\description{
`hclust_heatmap` takes a data frame and prints a heatmap from hierarchical clustering of the numeric variables
}
\details{
This function takes a data frame and transforms it into a matrix. The given matrix is then used to print a heatmap with hierarchical clustering.
}
\examples{
hclust_heatmap(USArrests)
hclust_heatmap(iris, row_names = iris$Species, dist_method = "manhattan", hclust_method = "average")
}
