% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_clustering.R
\name{hclust_heatmap_interactive}
\alias{hclust_heatmap_interactive}
\title{Interactive hclust heatmap}
\usage{
hclust_heatmap_interactive(
  input_df,
  row_names = rownames(input_df),
  col_names = colnames(dplyr::select_if(input_df, is.numeric)),
  dist_method = "euclidean",
  hclust_method = "complete",
  title = "Hierarchical clustering",
  out_path = "none"
)
}
\arguments{
\item{input_df}{data frame.}

\item{row_names}{data frame. Default are the row names of the input data frame.}

\item{col_names}{data frame. Default are the column names of the input data frame.}

\item{dist_method}{string. Method for the dist() function. dist_method = c("euclidean", "manhattan"). Default = "euclidean".}

\item{hclust_method}{string. Method for the hclust() function. hclust_method = c("complete", "single", "average"). Default = "complete".}

\item{title}{string. Plot title. Default = "Hierarchical clustering".}

\item{out_path}{string. Path to save plot to png.}
}
\description{
`hclust_heatmap_interactive` takes a data frame and lets you open an interactive heatmap in the viewer or save the heatmap to png.
}
\details{
This finction takes a data frame (optionally numeric or numeric with additional non-numeic varibles). Non-numeric variables are not part of the heatmap, but displayed in a separate box where the levels are color coded. The used method to create the heatmap is hierarchical clustering. Dendrograms are displayed for bowth rows and columns.
}
\examples{
hclust_heatmap_interactive(iris, iris$Species)
hclust_heatmap_interactive(USArrests,
                           col_names = c("M", "A", "UP", "R"),
                           row_names = 1:50,
                           title = "Heatmap of USArrests")
\dontrun{
dir <- "examples/USArrests"
hclust_heatmap_interactive(USArrests, out_path = dir)}
}
