% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_clustering.R
\name{hclust_methods}
\alias{hclust_methods}
\title{Hierarchical clustering method comparison}
\usage{
hclust_methods(
  input_df,
  labs = row.names(input_df),
  title = "Hierarchical clustering method comparison",
  dist_method = "euclidean",
  out_path = "none"
)
}
\arguments{
\item{input_df}{data frame.}

\item{labs}{vector. Labels of dendrogram. Default: rownames(input_df)}

\item{title}{string. Plot title. Default = "Hierarchical clustering method comparison".}

\item{dist_method}{string. MEthod for distance function. = c("euclidean", "manhattan")}

\item{out_path}{string. Path to save plot to png. Default = to device.}
}
\description{
`hclust_methods` takes a data frame and prints three dendrograms, one each for average linkage, single linkage and complete linkage.
}
\details{
This function takes a data frame and transforms it into a matrix. The given matrix is then used to perform hiererchical clustering with the average, single and complete linkage methods. The three clustering methods are then displayed as dendrograms. This plot is ment as a help to find ozt which clustering method should be used.
}
