% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_clustering.R
\name{hclust_performance_plot}
\alias{hclust_performance_plot}
\title{Hierarchical clustering performance plot}
\usage{
hclust_performance_plot(
  input_df,
  dist_methods = c("euclidean", "manhattan"),
  hclust_methods = c("average", "single", "complete"),
  out_path = "none"
)
}
\arguments{
\item{input_df}{data frame.}

\item{dist_methods}{vector. Options: c("euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski"). Default: c("euclidean", "maximum", "manhattan", "canberra", "minkowski")}

\item{hclust_methods}{vector. Options: c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid"). Default: c("average", "single", "complete")}

\item{out_path}{string. Path to save the text output to. If specified, the output is saved in a .txt-file. If not specified, the output is printed th the device.}
}
\description{
`hclust_performance_plot` takes a data frame and calculates which distance function and hclust function combination has the best performance.
}
\details{
This function calculated the performance for all combinations of the following distance and hclust functions distance functions = c("euclidean", "manhattan"), hclust functions = c("average", "single", "complete")). The results are printed in a table. The combination with the highest performance value will produce the best clustering.
}
