% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_visualization.R
\name{histogram_by_factor}
\alias{histogram_by_factor}
\title{Print one histogram for each variable group.}
\usage{
histogram_by_factor(input_df, factor, out_path = "none")
}
\arguments{
\item{input_df}{a data frame with at least one factor column}

\item{factor}{a string with the column name to group by}

\item{out_path}{optional string.}
}
\description{
`histogram_by_factor` prints a histogram with density line for each group and variable of a data frame
}
\details{
Generate a new grid/pdf dochument for each group of a data frame. Generate one plot for each variable of this subset data frame.
}
\examples{
histogram_by_factor(iris, "Species")
\dontrun{
dir.create(paste(getwd(), "/examples", sep = ""), showWarnings = FALSE)
dir <- paste(getwd(), "/examples/iris", sep = "")
histogram_by_factor(iris, "Species", dir)
}
}
