% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_hypothesis_testing.R
\name{list_anova_by_col}
\alias{list_anova_by_col}
\title{ANOVA list}
\usage{
list_anova_by_col(input_df, factor_col, print_all = FALSE)
}
\arguments{
\item{input_df}{data frame. Has multiple columns with numerical variables and at least one column with a factor variable.}

\item{factor_col}{string. Gives the name of the column, where the groups between which the anova is performed are stored.}

\item{print_all}{bool. Prints a summary of each anova. Default = FALSE}
}
\description{
`one_way_anova_by_col` performs an ANOVA for each numeric column of a data frame and returns f-statistic and p-value in a data frame.
}
