% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_hypothesis_testing.R
\name{log2_foldchange}
\alias{log2_foldchange}
\title{Calculate log2 foldchange}
\usage{
log2_foldchange(input_df, group_vector, control_group, test_group)
}
\arguments{
\item{input_df}{data frame. Contains at least one numeric variable.}

\item{group_vector}{vector. Vctor containing the grouping variabe. group_vector = input_df$group (or any vector of length nrow(input_df))}

\item{control_group}{string. Name of control group.}

\item{test_group}{string. Name of test group.}
}
\description{
`log2_foldchange` calculates the log2 foldchange for two groups of a given data frame.
}
\details{
This function takes a data frame and a grouping vector and calculates the log2 foldchange between two groups.
The results are returned in a new data frame.
}
