% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_visualization.R
\name{matrix_heatmap}
\alias{matrix_heatmap}
\title{Matrix heatmap}
\usage{
matrix_heatmap(input_df, title = "", interactive = FALSE, out_path = "none")
}
\arguments{
\item{input_df}{data frame or matrix. For exaple correlation or matrix of ratios or differences.}

\item{title}{string. Main title of the plot. Default: no title}

\item{interactive}{logical. Print heatmap to device (FALSE, default) or open interactive heatmap in browser (TRUE).}

\item{out_path}{string. Path to save heatmap to png. If out_path is empty the heatmap is printed to the device.}
}
\description{
`matrix_heatmap` displays values of a matrix in a heatmap
}
\examples{
iris_cor <- cor(iris[,-5], method = "spearman")
matrix_heatmap(iris_cor)
\dontrun{
dir.create(paste(getwd(), "/examples", sep = ""), showWarnings = FALSE)
dir <- paste(getwd(), "/examples/iris", sep = "")
matrix_heatmap(iris_cor, interactive = TRUE)
matrix_heatmap(iris_cor, out_path = dir)
}
}
