% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Motor Trend Car Road Tests from R datasets}
\format{
An object of class \code{data.frame} with 32 rows and 11 columns.
}
\usage{
mtcars
}
\description{
Motor Trend Car Road Tests from R datasets
}
\keyword{datasets}
