% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_hypothesis_testing.R
\name{one_sample_test_by_col}
\alias{one_sample_test_by_col}
\title{Perform one sample test by column}
\usage{
one_sample_test_by_col(
  input_df,
  group_vector,
  alternative = "two.sided",
  method = t.test,
  confidence_level = 0.95
)
}
\arguments{
\item{input_df}{data frame. Contains at least one numeric variable.}

\item{group_vector}{vector. Vector containing the groups to be compared. group_vector = input_df$group (or other vector of length nrow(input_df))}

\item{alternative}{string. Specifies the alternative hypothesis. c("two.sided" (default), "greater", "less")}

\item{method}{function. Options: t.test (default), wilcox.test}

\item{confidence_level}{numeric. Confidence level of the interval. Default = 0.95}
}
\description{
`p_values_by_column` takes a data frame and a grouping vector and returns a data frame with the p-values of a given method.
}
\details{
This function takes a data frame ans a grouping vector and performs a given hypothesis test on all numeric variables. The results are returned in a new data frame.
}
\examples{
my_iris <- subset(x = iris, Species == "setosa" | Species == "versicolor")
one_sample_test_by_col(my_iris, my_iris$Species)
one_sample_test_by_col(my_iris, my_iris$Species, method = wilcox.test)
}
