% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_pca.R
\name{plot_loadings}
\alias{plot_loadings}
\title{Loadings plot}
\usage{
plot_loadings(
  pca_object,
  colour = FALSE,
  top_loadings = 10,
  title = "Loadings plot",
  xlab = "Dim.1",
  ylab = "Dim.2"
)
}
\arguments{
\item{pca_object}{Object produced by FactoMineR::PCA.}

\item{colour}{logical. TRUE ... color scale by PC1. FALSE (default) ... all black.}

\item{top_loadings}{integer. Number of loadings to label (the top n loadings of both PCs are labelled).}

\item{title}{string. Main title of the plot. Default = "Loadings plot"}

\item{xlab}{string. Title of x-axis. Default: "Dim.1"}

\item{ylab}{string. Title of y-axis. Default: "Dim.2"}
}
\description{
`plot_loadings` plots the loadings between the first two principal components.
}
