% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_pca.R
\name{plot_pc_variance}
\alias{plot_pc_variance}
\title{Plot contribution of principal component to variance}
\usage{
plot_pc_variance(
  data_frame,
  x,
  y,
  xlab = "dimensions",
  ylab = "variance [\%]",
  title = "Scree plot",
  fill = "#35608DFF",
  hjust = -0.2
)
}
\arguments{
\item{data_frame}{data frame. Contains eigenvalues, contribution to data variance and cummulative contribution to data variance for a PCA model.}

\item{x}{column of data_frame. Column with principal component names.}

\item{y}{column of data_frame. Column with eigenvalues/contribution to variance/ cummulative contribution to variance.}

\item{xlab}{string. X-axis label. Default: "dimensions"}

\item{ylab}{string. X-axis label. Default: "variance [\%]"}

\item{title}{string. Main title. Default: "Scree plot"}

\item{fill}{string. Color of the barplot. Default: "#35608DFF"}

\item{hjust}{numeric. Position of labels relative to the points in the graph. Default: -0.2}
}
\description{
`plot_pc_variance` prints a barplot with a line of the contribution of principal components to the data variance
}
\examples{
{iris_pca <- FactoMineR::PCA(dplyr::select_if(iris, is.numeric))
iris_eigen <- as.data.frame(factoextra::get_eigenvalue(iris_pca))
plot_pc_variance(iris_eigen, x = rownames(iris_eigen), y = iris_eigen$cumulative.variance.percent)}
}
