% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_plsDA.R
\name{plot_pls_scores}
\alias{plot_pls_scores}
\title{PLS-DA scores plot}
\usage{
plot_pls_scores(
  pls_da_obj,
  colour = TRUE,
  top_loadings = 10,
  title = "PLS-DA",
  xlab = "PLS-DA 1",
  ylab = "PLS-DA 2"
)
}
\arguments{
\item{pls_da_obj}{Object produced by MixOmics::splsda or MixOmics::plsda.}

\item{colour}{logical. TRUE ... color scale by PC1. FALSE (default) ... all black.}

\item{top_loadings}{integer. Number of loadings to label (the top n loadings of both PCs are labelled).}

\item{title}{string. Main title of the plot. Default = "PLS-DA"}

\item{xlab}{string. Title of x-axis. Default: "PLS-DA 1"}

\item{ylab}{string. Title of y-axis. Default: "PLS-DA 2"}
}
\description{
`plot_pls_scores` plots the loadings between the first two principal components.
}
