% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_visualization.R
\name{plot_ratio_barcharts}
\alias{plot_ratio_barcharts}
\title{Plot ratio barcharts}
\usage{
plot_ratio_barcharts(data_frame, subset_vector, group_column)
}
\arguments{
\item{data_frame}{data frame. From aggregate() function.}

\item{subset_vector}{vector of column names from the data_frame.}

\item{grop_column}{string. Name of the column to group by.}
}
\description{
`plot_ratio_barcharts` plots the ratios of a list of values from a data frame by group.
}
