% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_read_TF.R
\name{pretty_transpose}
\alias{pretty_transpose}
\title{Transpose data frame with correct row- and column names.}
\usage{
pretty_transpose(input_df)
}
\arguments{
\item{input_df}{a data frame with character strings and numerics}
}
\description{
`pretty_transpose` transposes a data frame to a new data frame
}
\details{
This function takes a data frame, transposes it and sets the former first column as the column names.
Row names are automatically generated by the data.table::transpose() function.
}
\examples{
df <- as.data.frame(cbind(c("ID", "a", "b", "c", "d"), c("z", 1, 2, 3, 4), c("x", 5, 6, 7, 8)))
pretty_transpose(df)
}
