% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_visualization.R
\name{qqplot_by_factor}
\alias{qqplot_by_factor}
\title{Print one qqplot for each variable and group}
\usage{
qqplot_by_factor(input_df, factor, out_path = "none")
}
\arguments{
\item{input_df}{a data frame with at least one factor column.}

\item{factor}{a string with the column name to group by}

\item{out_path}{optional string. If out path is given, the generated plots are saved to a pdf document with the given name.}
}
\description{
`qqplot_by_factor` takes a data frame and prints a qq-plot for each group and varible
}
\details{
Take a data frame and prints a qq-plot for each group and varible.
}
\examples{
qqplot_by_factor(iris, "Species")
\dontrun{
dir.create(paste(getwd(), "/examples", sep = ""), showWarnings = FALSE)
dir <- paste(getwd(), "/examples/iris", sep = "")
qqplot_by_factor(iris, "Species", dir)
}
}
