% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_pca.R
\name{scree_base}
\alias{scree_base}
\title{Scree plot with Rbase}
\usage{
scree_base(
  input_df,
  title = c("Scree plot", "Cummulative scree plot"),
  scale = FALSE,
  out_path = "none"
)
}
\arguments{
\item{input_df}{data frame or matrix.}

\item{title}{vector of strings. Main titles for both plots. Default: c("Scree plot", "Cummulative scree plot")}

\item{scale}{bool. Apply scaling when erforming PCA. Default = FALSE.}

\item{out_path}{string. Path to save plot to png. If out_path is empty, the plot is printed to the device.}
}
\description{
`scree_base` takes a prcomp element and diplays the percentage and the cummulative percentage of variance of variance explained by the principal components in a barplot.
}
\details{
This function takes a prcomp element (i.e. an element generated by the base::prcomp function), and displays the percentage of variance and the cummulative percentage of variance explained by each principal component in a dot- and line graph.
}
\examples{
scree_base(mtcars)
}
