% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_pca.R
\name{scree_factoextra}
\alias{scree_factoextra}
\title{Scree plot with factoextra}
\usage{
scree_factoextra(prcomp_element, title = "Scree plot", out_path = "none")
}
\arguments{
\item{prcomp_element}{object of class prcomp or pca. Produced by base::prcomp or FactoMineR::PCA.}

\item{title}{string. Main title of plot. Default: "Scree plot"}

\item{out_path}{string. Path to save plot to png. If out_path is empty, the plot is printed to the device.}
}
\description{
`scree_factoextra` takes a prcomp element and diplays the percentage of variance explained by the principal components in a barplot.
}
\details{
This function takes a prcom element (i.e. an element generated by the base::prcomp function), and displays the percentage of variance explained by each principal component in a barplot. Additionally there is also a dot- and line graph.
}
\examples{
{pca_iris <- prcomp(dplyr::select_if(iris, is.numeric))
scree_factoextra(pca_iris, title = "Scree plot iris")}
}
