% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_visualization.R
\name{simple_barchart}
\alias{simple_barchart}
\title{Simple barchart}
\usage{
simple_barchart(
  data_frame,
  x,
  y,
  fill = "#35608DFF",
  title = "test",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{data_frame}{data frame with at least one numeric and one factor column}

\item{x}{vector. Factor column}

\item{y}{vector. Numeric column}

\item{fill}{vector. Factor column to fill bars by. By default all bars are blue.}

\item{title}{string. Main title of the plot. Default = ""}

\item{xlab}{string. Title of x-axis. Default = ""}

\item{ylab}{string. Title of y-axis. Default = ""}
}
\description{
`simple_barchart` returns a ggplot barchart with viridis coloring
}
