% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tfData}
\alias{tfData}
\title{TraceFinder example data}
\format{

}
\usage{
tfData
}
\description{
A dataset containing LC-MS data after compound identification with Thermo Fisher Tracefinder.
}
\keyword{datasets}
