% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USArrests}
\alias{USArrests}
\title{Violent Crime Rates by US State from R datasets}
\format{
An object of class \code{data.frame} with 50 rows and 4 columns.
}
\usage{
USArrests
}
\description{
Violent Crime Rates by US State from R datasets
}
\keyword{datasets}
