% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidome_comparison_hypothesis_testing.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot}
\usage{
volcano_plot(
  volcano_df,
  foldchange_col,
  significance_col,
  significance = 0.05,
  foldchange = 1,
  title = "Volcano plot",
  x_lab = "log2Foldchange",
  y_lab = "-log10(p-value)",
  labels = TRUE,
  out_path = "none"
)
}
\arguments{
\item{volcano_df}{data frame.}

\item{foldchange_col}{vector. Column of volcano_df containing the foldchange values. Format: volcano_df$foldchange_col}

\item{significance_col}{vector. Column of volcano_df containing the p-values. Format: volcano_df$significance_col}

\item{significance}{numeric. Value for the significance threshold. Default = 0.05.}

\item{foldchange}{numeric. Vaue for the foldchange threshold. Default = #todo}

\item{title}{string. Main title. Default = "Volcano plot"}

\item{x_lab}{string. x-axix title. Default = "log2Fold"}

\item{y_lab}{string. y-axis title- Default = "-log10(p-value)"}

\item{labels}{logical. Should the points over the threshold be labelled. Options: TRUE (default), FALSE}

\item{out_path}{string. Path to save plot to png. If out_path is empty, the plot is printed to the device. If "none", the plot is either printed or saved to a variabe.}
}
\description{
`volcano_plot` prints a volcano plot using a data frame containing p-values and foldchanges
}
\details{
This function takes a data frame with columns for p-values and (log2)foldchanges and returns a dotplot resembelling a volcano plot.
It uses the given significance and foldchange threshods to mark significantly up- and down-regulated elements by color and by displaying
threshold lines.
}
